pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sushi"
version = "0.1.0"
description = "A collection of shared Python and R tools for the sushi project." # Updated description
authors = [
  { name="Your Name", email="you@example.com" }, # Update with your details
]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Private :: Do Not Upload",
]

# Merged list of all your Python dependencies
dependencies = [
    "boto3>=1.40.51",
    "botocore>=1.40.51",
    "pandas>=2.3.3",
    "polars>=1.34.0",
    "pyyaml>=6.0.3",
    "google-cloud-bigquery",
    "pyfarmhash",
    "requests",
]

[project.urls]
"Homepage" = "https://github.com/your-org/sushi" # Example URL

# --- CRUCIAL SECTION FOR INSTALLATION ---
[tool.setuptools]
# Tells setuptools that your Python package code is inside the 'sushi' directory
packages = ["sushi"]

[tool.setuptools.package-data]
# This ensures that schema.yaml is included when the package is installed
sushi = ["utils/io/schema.yaml"]

